Update to 1.19.1-pre6
We had to nullify signing keys from the player list packet due to a change in this snapshot, that will disconnect the client if there is a known signing key for a user but that user sends unsigned message. Sadly, it means a re-login after switching to STRIP_SIGNATURE will be needed (or all players will disconnect).